requirements.txt
openai-whisper
python-docx